# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready blog backend API for Alfafaa Community - a Muslim community blogging platform.",
        "title": "Alfafaa Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alfafaa Support",
            "url": "https://alfafaa.com/support",
            "email": "support@alfafaa.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/media": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of all media files (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "List all media",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.MediaResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires admin role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/media/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a file (image only, requires authentication)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Upload media",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alt text for the image",
                        "name": "alt_text",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.MediaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "No file provided or invalid file type",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/media/{id}": {
            "get": {
                "description": "Get a single media file info by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Get media by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.MediaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Media not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a media file (owner or admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Delete media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not the owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Media not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.MediaResponse": {
            "type": "object",
            "properties": {
                "alt_text": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_size_formatted": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "original_filename": {
                    "type": "string"
                },
                "uploaded_by": {
                    "type": "string"
                },
                "uploader": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.PublicUserResponse"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_dto.PublicUserResponse": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_utils.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}